+++ crates/sui-json-rpc/src/lib.rs
@@ -50,6 +50,9 @@ pub const MAX_REQUEST_SIZE: u32 = 2 << 30;
 #[cfg(test)]
 #[path = "unit_tests/rpc_server_tests.rs"]
 mod rpc_server_test;
+#[cfg(test)]
+#[path = "unit_tests/transaction_tests.rs"]
+mod transaction_tests;
 
 pub struct JsonRpcServerBuilder {
     module: RpcModule<()>,
